+++ util.go
@@ -84,17 +84,12 @@ package u2f
 import (
 	"crypto/rand"
 	"crypto/subtle"
-    "crypto/elliptic"
-    "crypto/x509"
 	"encoding/base64"
 	"encoding/json"
 	"errors"
 	"strings"
 	"time"
-    "fmt"
 )
-
-import "reflect"
 
 const u2fVersion = "U2F_V2"
 const timeout = 5 * time.Minute
@@ -141,159 +136,6 @@ func NewChallenge(appID string, trustedFacets []string, registeredKeys []Registr
 	c.RegisteredKeys = registeredKeys
 
 	return &c, nil
-}
-
-
-func reflectSetFieldString(i interface{}, name string, value string) error {
-    r := reflect.ValueOf(i).Elem()
-
-    field := r.FieldByName(name)
-    if field.Kind() == reflect.Invalid {
-        return fmt.Errorf("Cannot find field: %s", name)
-    }
-    if field.Kind() != reflect.String {
-        return fmt.Errorf("Invalid field type: %s field: %s", field.Kind(), name)
-    }
-    if !field.CanSet() {
-        return fmt.Errorf("Cannot set field: %s", name)
-    }
-    field.SetString(value)
-
-    return nil
-}
-
-func reflectGetFieldString(i interface{}, name string) (string, error) {
-    r := reflect.ValueOf(i).Elem()
-
-    field := r.FieldByName(name)
-    if field.Kind() == reflect.Invalid {
-        return "", fmt.Errorf("Cannot find field: %s", name)
-    }
-    if field.Kind() != reflect.String {
-        return "", fmt.Errorf("Invalid field type: %s field: %s", field.Kind(), name)
-    }
-
-    return field.String(), nil
-}
-
-func reflectSetFieldUint(i interface{}, name string, value uint32) error {
-    r := reflect.ValueOf(i).Elem()
-
-    field := r.FieldByName(name)
-    if field.Kind() == reflect.Invalid {
-        return fmt.Errorf("Cannot find field: %s", name)
-    }
-    if field.Kind() != reflect.Uint {
-        return fmt.Errorf("Invalid field type: %s field: %s", field.Kind(), name)
-    }
-    if !field.CanSet() {
-        return fmt.Errorf("Cannot set field: %s", name)
-    }
-    field.SetUint(uint64(value))
-
-    return nil
-}
-
-func reflectGetFieldUint(i interface{}, name string) (uint, error) {
-    r := reflect.ValueOf(i).Elem()
-
-    field := r.FieldByName(name)
-    if field.Kind() == reflect.Invalid {
-        return 0, fmt.Errorf("Cannot find field: %s", name)
-    }
-    if field.Kind() != reflect.Uint {
-        return 0, fmt.Errorf("Invalid field type: %s field: %s", field.Kind(), name)
-    }
-
-    return uint(field.Uint()), nil
-}
-
-func (reg *Registration) MarshalStruct(i interface{}) error {
-    r := reflect.ValueOf(i).Elem()
-
-    // Check the interface is a struct
-    if r.Kind() != reflect.Struct {
-        return fmt.Errorf("Interface is not a struct (type is: %s)", r.Kind())
-    }
-
-    err := reflectSetFieldString(i, "KeyHandle", string(reg.KeyHandle))
-    if err != nil {
-        return err
-    }
-
-    publicKeyString := encodeBase64(elliptic.Marshal(reg.PubKey.Curve, reg.PubKey.X, reg.PubKey.Y))
-    err = reflectSetFieldString(i, "PublicKey", publicKeyString)
-    if err != nil {
-        return err
-    }
-
-    certString := encodeBase64(reg.AttestationCert.Raw)
-    err = reflectSetFieldString(i, "Certificate", certString)
-    if err != nil {
-        return err
-    }
-
-    err = reflectSetFieldUint(i, "Counter", reg.Counter)
-    if err != nil {
-        return err
-    }
-
-    return err
-}
-
-func (reg *Registration) UnmarshalStruct(i interface{}) error {
-    r := reflect.ValueOf(i).Elem()
-
-    // Check the interface is a struct
-    if r.Kind() != reflect.Struct {
-        return fmt.Errorf("Interface is not a struct (type is: %s)", r.Kind())
-    }
-
-    // Key handle
-    keyHandleString, err := reflectGetFieldString(i, "KeyHandle")
-    if err != nil {
-        return err
-    }
-    reg.KeyHandle = []byte(keyHandleString)
-
-    // Public key
-    publicKeyString, err := reflectGetFieldString(i, "PublicKey")
-    if err != nil {
-        return err
-    }
-
-    publicKeyDecoded, err := decodeBase64(publicKeyString)
-    if err != nil {
-        return err
-    }
-
-    reg.PubKey.X, reg.PubKey.Y = elliptic.Unmarshal(elliptic.P256(), publicKeyDecoded)
-    reg.PubKey.Curve = elliptic.P256()
-
-    // Attestation certificate
-    certString, err := reflectGetFieldString(i, "Certificate")
-    if err != nil {
-        return err
-    }
-    certStringDecoded, err := decodeBase64(certString)
-    if err != nil {
-        return err
-    }
-    cert, err := x509.ParseCertificate(certStringDecoded)
-    if err != nil {
-        return err
-    }
-
-    reg.AttestationCert = cert
-
-    Counter, err := reflectGetFieldUint(i, "Counter")
-    if err != nil {
-        return err
-    }
-
-    reg.Counter = uint32(Counter)
-
-    return nil
 }
 
 func verifyClientData(clientData []byte, challenge Challenge) error {
